requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
astroid==3.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
dill==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-stubs==3.14.5
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isort==5.13.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
Twisted==23.10.0
txaio==23.1.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.4
zope.interface==6.1
